pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-to-policy"
version = "0.0.1"
description = "Educational demo: natural-language prompts to trained RL policies via LLM-emitted reward specs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Finding Theta" }]
dependencies = [
    "gymnasium[mujoco]>=1.0",
    "stable-baselines3>=2.3",
    "anthropic>=0.40",
    "imageio>=2.34",
    "imageio-ffmpeg>=0.5",
    "pydantic>=2.7",
    "tyro>=0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]
# Reward generation via Google Gemini. Adds the ``google-genai`` SDK; the
# Anthropic provider remains available without this extra installed.
gemini = [
    "google-genai>=0.3",
]
# Run the reward LLM locally (e.g. on a Colab GPU) instead of calling an
# API. ``bitsandbytes`` is the 4-/8-bit quantizer that lets 7-8B models
# fit on a T4 (and lets larger models fit on an A100 / L4).
# ``accelerate`` is required by transformers' ``device_map="auto"``.
# ``torch`` is already pulled in by SB3.
local = [
    "transformers>=4.45",
    "accelerate>=0.33",
    "bitsandbytes>=0.43",
]
# Convenience: install everything (all providers + dev tools).
all = [
    "prompt-to-policy[dev,gemini,local]",
]

[project.scripts]
p2p = "prompt_to_policy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_to_policy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]

[tool.ruff.lint.per-file-ignores]
# The prompt templates' few-shot examples are intentionally on single
# lines so that the rendered LLM prompt matches what we tested.
"src/prompt_to_policy/llm/templates/halfcheetah.py" = ["E501"]
"src/prompt_to_policy/llm/templates/hopper.py" = ["E501"]
"src/prompt_to_policy/llm/templates/ant.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"
markers = [
    "slow: tests that take more than a few seconds (training harness smoke).",
]
